requirements.txt.txt
ai==0.0.3
arrow==0.15.4
astroid==2.2.5
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
beautifulsoup4==4.8.2
boto3==1.11.3
botocore==1.14.3
build==1.0.2
CacheControl==0.12.6
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
construct==2.9.47
cycler==0.10.0
django-crispy-forms==1.7.2
djangoadmin==1.0.9
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
Flask==1.1.1
future==0.17.1
genomes==1.1.0
geos==0.2.2
gunicorn==20.0.4
gis==0.2.1
google==2.0.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.21
incremental==17.5.0
iso3166==1.0.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
leaflet==0.0.3
lxml==4.4.2
mapbox==0.18.0
MarkupSafe==1.1.1
matlab==0.1
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
numpy==1.17.4
pandas==0.25.3
parso==0.5.1
Pillow==6.1.0
pipenv==2018.11.26
pluggy==0.13.0
ply==3.11
polyline==1.4.0
postgis==1.0.4
postgres==3.0.0
proj==0.1.0
protobuf==3.11.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycryptodomex==3.9.4
pydocstyle==4.0.1
pyflakes==2.1.1
PyHamcrest==1.10.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.6
pypi==2.1
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-jsonrpc-server==0.2.0
python-language-server==0.28.3
pytz==2019.2
PyYAML==5.3
requests==2.22.0
rope==0.14.0
s3transfer==0.3.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
six==1.12.0
snowballstemmer==1.9.1
soupsieve==1.9.5
spatial==0.2.0
spatialite==0.0.3
sqlparse==0.3.0
stripe==2.33.0
tensor==0.3.6
tracker==0.1.1
Twisted==19.10.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.3
virtualenv==16.7.2
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
zope.interface==4.7.1
